+++ core/src/test/java/com/bobo/storage/core/service/impl/PlaylistSongServiceImplTest.java
@@ -8,6 +8,7 @@ import com.bobo.storage.core.domain.PlaylistSongMother;
 import com.bobo.storage.core.resource.access.PlaylistSongRepository;
 import com.bobo.storage.core.resource.query.PlaylistSongQueryRepository;
 import com.bobo.storage.core.service.PlaylistSongService;
+import com.bobo.storage.core.service.SongService;
 import java.util.Random;
 import org.junit.jupiter.api.*;
 
@@ -27,7 +28,9 @@ class PlaylistSongServiceImplTest {
 	@BeforeEach
 	void setUp() {
 		repository = mock(PlaylistSongRepository.class);
-		service = new PlaylistSongServiceImpl(repository, mock(PlaylistSongQueryRepository.class));
+		service =
+				new PlaylistSongServiceImpl(
+						repository, mock(PlaylistSongQueryRepository.class), mock(SongService.class));
 	}
 
 	/**
